Moved ProfileView to features directory
